Use uint64 for indirect pin refcounts
Platform-dependent behavior is not nice, and negative refcounts are
not very useful.

+++ core/commands/pin.go
@@ -273,7 +273,7 @@ Defaults to "direct".
 
 type RefKeyObject struct {
 	Type  string
-	Count int
+	Count uint64
 }
 
 type RefKeyList struct {
